requirements.txt
knackpy==1.0.*
mstrio-py==11.3.4.*
pandas
boto3==1.19.*
sodapy==2.1.*
smartsheet-python-sdk==3.0.*
pyproj
pytz
openpyxl
oracledb==2.1.*
python-dateutil==2.8.*
